requirements.txt
asgiref==3.5.2
attrs==22.2.0
boto3==1.24.86
botocore==1.27.86
Django==4.1.1
django-cors-headers==3.13.0
django-enumfields==2.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
exceptiongroup==1.1.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.4
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
